refactor: name checker files now lives in a dedicated folder
