+++ custom_lint_brick/__brick__/mason_custom_lints/src/name_checkers/{{#arrayValues}}{{{valueName}}}_name_checker.dart{{/arrayValues}}
@@ -5,7 +5,7 @@ import 'package:analyzer_plugin/protocol/protocol_generated.dart';
 import 'package:analyzer_plugin/utilities/change_builder/change_builder_core.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
-import 'extensions.dart';
+import '../extensions.dart';
 
 Stream<Lint> {{arrayValues.valueName}}ShouldMatchFileName({
     required ResolvedUnitResult unit,
