+++ mason_custom_lints/bin/custom_lint.dart
@@ -4,9 +4,9 @@ import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
-import '../src/service_name_checker.dart';
+import '../src/name_checkers/service_name_checker.dart';
 
-import '../src/bloc_name_checker.dart';
+import '../src/name_checkers/bloc_name_checker.dart';
 
 void main(List<String> args, SendPort sendPort) {
   startPlugin(sendPort, _MasonCustomLints());
